3장 예제 10번까지(result1 정상적인 값으로 출력시키는 방법)

+++ ch03/src/ch03/OperatorEx10.java
@@ -4,7 +4,7 @@ public class OperatorEx10 {
 	public static void main(String[] args) {
 		int a = 1000000;
 		
-		int result1 = a * a / a;	// 1000000 * 1000000 / 1000000
+		long result1 = (long)a * a / a;	// 1000000 * 1000000 / 1000000
 		int result2 = a / a * a;	// 1000000 / 1000000 * 1000000
 		
 		System.out.printf("%d * %d / %d = %d%n", a, a, a, result1);
